Remove the USB features as they are now located at https://github.com/wisespace-io/yubico-manager
