+++ examples/README.md
@@ -11,23 +11,3 @@ cargo run --release --example "otp_custom"
 ## Asynchronous OTP check
 
 cargo run --release --example "otp_async"
-
-## Device Configuration (HMAC)
-
-cargo run --release --example "configuration_hmac"
-
-## Challenge-Response (HMAC)
-
-cargo run --release --example "challenge_response_hmac"
-
-## Device Configuration (OTP)
-
-cargo run --release --example "configuration_otp"
-
-## Challenge-Response (OTP)
-
-cargo run --release --example "challenge_response_otp"
-
-## Read Serial Number
-
-cargo run --release --example "serial_number"
