--- examples/challenge_response_hmac.rs
@@ -1,35 +1,0 @@
-extern crate hex;
-extern crate yubico;
-
-use std::ops::Deref;
-use yubico::{Yubico};
-use yubico::config::{Config, Slot, Mode};
-
-fn main() {
-   let mut yubi = Yubico::new();
-
-   if let Ok(device) = yubi.find_yubikey() {
-       println!("Vendor ID: {:?} Product ID {:?}", device.vendor_id, device.product_id);
-
-       let config = Config::default()
-           .set_vendor_id(device.vendor_id)
-           .set_product_id(device.product_id)
-           .set_variable_size(true)
-           .set_mode(Mode::Sha1)
-           .set_slot(Slot::Slot2);
-
-       // Challenge can not be greater than 64 bytes
-       let challenge = String::from("mychallenge");
-       // In HMAC Mode, the result will always be the SAME for the SAME provided challenge
-       let hmac_result= yubi.challenge_response_hmac(challenge.as_bytes(), config).unwrap();
-
-       // Just for debug, lets check the hex
-       let v: &[u8] = hmac_result.deref();
-       let hex_string = hex::encode(v);
-
-       println!("{}", hex_string);       
-     
-   } else {
-       println!("Yubikey not found");
-   }
-}
