+++ src/config.rs
@@ -50,32 +50,10 @@ impl Display for SyncLevel{
     }
 }
 
-#[derive(Clone, Copy, Debug, PartialEq)]
-#[repr(u8)]
-pub enum Command {
-    Configuration1 = 0x01,
-    Configuration2 = 0x03,
-    Update1 = 0x04,
-    Update2 = 0x05,
-    Swap = 0x06,
-    DeviceSerial = 0x10,
-    DeviceConfig = 0x11,       
-    ChallengeOtp1 = 0x20,
-    ChallengeOtp2 = 0x28,
-    ChallengeHmac1 = 0x30,
-    ChallengeHmac2 = 0x38,
-}
-
 #[derive(Clone, Debug, PartialEq)]
 pub struct Config {
     pub client_id: String,
     pub key: Vec<u8>,
-    pub product_id: u16,
-    pub vendor_id: u16,
-    pub variable: bool,
-    pub slot: Slot,
-    pub mode: Mode,
-    pub command: Command,
     pub api_hosts: Vec<String>,
     pub user_agent: String,
     pub sync_level: SyncLevel,
@@ -87,19 +65,13 @@ impl Config {
         Config {
             client_id: String::new(),
             key: Vec::new(),
-            product_id: 0x00,
-            vendor_id: 0x1050,                       
-            variable: true,
-            slot: Slot::Slot1,
-            mode: Mode::Sha1,
-            command: Command::ChallengeHmac1,
             api_hosts: build_hosts(),
             user_agent: "github.com/wisespace-io/yubico-rs".to_string(),
             sync_level: SyncLevel::secure(),
         }
     }
 
-    pub fn set_client_id<C>(mut self, client_id: C) -> Self 
+    pub fn set_client_id<C>(mut self, client_id: C) -> Self
         where C: Into<String>
     {
         self.client_id = client_id.into();
@@ -115,36 +87,6 @@ impl Config {
 
     pub fn set_api_hosts(mut self, hosts: Vec<String>) -> Self {
         self.api_hosts = hosts;
-        self
-    }
-
-    pub fn set_vendor_id(mut self, vendor_id: u16) -> Self {
-        self.vendor_id = vendor_id;
-        self
-    }
-
-    pub fn set_product_id(mut self, product_id: u16) -> Self {
-        self.product_id = product_id;
-        self
-    }
-    
-    pub fn set_variable_size(mut self, variable: bool) -> Self {
-        self.variable = variable;
-        self
-    }
-
-    pub fn set_slot(mut self, slot: Slot) -> Self {
-        self.slot = slot;
-        self
-    }
-
-    pub fn set_mode(mut self, mode: Mode) -> Self {
-        self.mode = mode;
-        self
-    }
-
-    pub fn set_command(mut self, command: Command) -> Self {
-        self.command = command;
         self
     }
